+++ src/main/java/com/java/poo/javapoo/domain/funcionario/model/Administrador.java
@@ -4,11 +4,12 @@ import com.java.poo.javapoo.domain.enums.Departamento;
 import com.java.poo.javapoo.domain.enums.TempoDeCasa;
 import com.java.poo.javapoo.domain.funcionario.Funcionario;
 
+import java.math.BigDecimal;
 
 
 public class Administrador extends Funcionario {
 
-    public Administrador(String nome, double salario, Departamento departamento, TempoDeCasa tempoDeCasa) {
+    public Administrador(String nome, BigDecimal salario, Departamento departamento, TempoDeCasa tempoDeCasa) {
         super(nome, salario, departamento, tempoDeCasa);
     }
 }
