+++ src/main/java/com/java/poo/javapoo/domain/funcionario/model/Design.java
@@ -4,10 +4,12 @@ import com.java.poo.javapoo.domain.enums.Departamento;
 import com.java.poo.javapoo.domain.enums.TempoDeCasa;
 import com.java.poo.javapoo.domain.funcionario.Funcionario;
 
+import java.math.BigDecimal;
+
 
 public class Design extends Funcionario {
 
-    public Design(String nome, double salario, Departamento departamento, TempoDeCasa tempoDeCasa) {
+    public Design(String nome, BigDecimal salario, Departamento departamento, TempoDeCasa tempoDeCasa) {
         super(nome, salario, departamento, tempoDeCasa);
     }
 }
