+++ src/test/java/com/java/poo/javapoo/domain/model/funcionario/ServicosGeraisTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.opentest4j.AssertionFailedError;
 
+import java.math.BigDecimal;
+
 public class ServicosGeraisTest {
 
     @Test
@@ -28,7 +30,7 @@ public class ServicosGeraisTest {
     }
 
     public static ServicosGerais getServicosGerais(){
-        var sg = new ServicosGerais("MAXWELL", 2222.00, Departamento.SERVICOS_GERAIS, TempoDeCasa.UM_ANO);
+        var sg = new ServicosGerais("MAXWELL", BigDecimal.valueOf(2222.00), Departamento.SERVICOS_GERAIS, TempoDeCasa.UM_ANO);
         return sg;
     }
 }
